requirements/base.txt
# Core
pytz==2018.9
Django==2.1.7
django-environ==0.4.5
gunicorn==19.9.0
newrelic==4.12.0.113

# For the persistence stores
psycopg2-binary==2.7.7
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.9.1
Markdown==3.0.1
django-filter==2.1.0

# Auth
djangorestframework-jwt==1.11.0
django-cors-headers==2.5.2
argon2-cffi==19.1.0

# Google Auth
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-oauthlib==0.3.0